requirements.txt
os-sys
tk
matplotlib
numpy
requests
pillow
python-dotenv